modify some public/private stuff and clean code
